@@ -28,51 +28,53 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.1.1",
-    "@angular/common": "13.1.1",
-    "@angular/compiler": "13.1.1",
-    "@angular/core": "13.1.1",
-    "@angular/forms": "13.1.1",
-    "@angular/platform-browser": "13.1.1",
-    "@angular/platform-browser-dynamic": "13.1.1",
-    "@angular/router": "13.1.1",
+    "@angular/animations": "14.0.0",
+    "@angular/common": "14.0.0",
+    "@angular/compiler": "14.0.0",
+    "@angular/core": "14.0.0",
+    "@angular/forms": "14.0.0",
+    "@angular/platform-browser": "14.0.0",
+    "@angular/platform-browser-dynamic": "14.0.0",
+    "@angular/router": "14.0.0",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
     "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.1.2",
+    "@angular-devkit/build-angular": "14.0.0",
     "@angular-eslint/eslint-plugin": "13.0.1",
     "@angular-eslint/eslint-plugin-template": "13.0.1",
     "@angular-eslint/template-parser": "13.0.1",
     "@angular-extensions/pretty-html-log": "^4.0.1",
-    "@angular/compiler-cli": "13.1.1",
-    "@angular/language-service": "13.1.1",
-    "@ngneat/spectator": "^7.2.0",
-    "@nrwl/angular": "13.4.3",
-    "@nrwl/cli": "13.4.3",
-    "@nrwl/eslint-plugin-nx": "13.4.3",
-    "@nrwl/jest": "13.4.3",
-    "@nrwl/linter": "13.4.3",
-    "@nrwl/tao": "13.4.3",
-    "@nrwl/workspace": "13.4.3",
-    "@types/jest": "27.0.2",
+    "@angular/compiler-cli": "14.0.0",
+    "@angular/language-service": "14.0.0",
+    "@ngneat/spectator": "^11.0.0",
+    "@nrwl/angular": "14.1.9",
+    "@nrwl/cli": "14.1.9",
+    "@nrwl/eslint-plugin-nx": "14.1.9",
+    "@nrwl/jest": "14.1.9",
+    "@nrwl/linter": "14.1.9",
+    "@nrwl/nx-cloud": "14.0.5",
+    "@nrwl/workspace": "14.1.9",
+    "@types/jest": "27.4.1",
     "@types/node": "14.14.33",
-    "@typescript-eslint/eslint-plugin": "5.3.1",
-    "@typescript-eslint/parser": "5.3.1",
+    "@typescript-eslint/eslint-plugin": "5.18.0",
+    "@typescript-eslint/parser": "5.18.0",
     "dotenv": "10.0.0",
-    "eslint": "8.2.0",
+    "eslint": "8.12.0",
     "eslint-config-prettier": "8.1.0",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.0.0",
-    "ng-packagr": "13.1.2",
+    "jest": "28.0.0",
+    "jest-environment-jsdom": "^28.1.1",
+    "jest-preset-angular": "12.1.0",
+    "ng-packagr": "14.0.0",
+    "nx": "14.1.9",
     "postcss": "^8.3.9",
-    "postcss-import": "^14.0.2",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-url": "^10.1.1",
-    "prettier": "2.5.1",
-    "ts-jest": "27.0.5",
+    "postcss-import": "14.1.0",
+    "postcss-preset-env": "7.5.0",
+    "postcss-url": "10.1.3",
+    "prettier": "2.6.2",
+    "ts-jest": "28.0.0",
     "ts-node": "~9.1.1",
-    "typescript": "4.5.4"
+    "typescript": "4.6.4"
   }
 }
