@@ -1,6 +1,6 @@
 {
   "name": "ngx-inject-control",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "Simple, reusable form components in Angular",
   "publishConfig": {
     "access": "public"
@@ -26,9 +26,9 @@
     "email": "jadengis@gmail.com"
   },
   "peerDependencies": {
-    "@angular/common": ">=12.0.0 <14.0.0",
-    "@angular/core": ">=12.0.0 <14.0.0",
-    "@angular/forms": ">=12.0.0 <14.0.0"
+    "@angular/common": ">=14.0.0 <15.0.0",
+    "@angular/core": ">=14.0.0 <15.0.0",
+    "@angular/forms": ">=14.0.0 <15.0.0"
   },
   "dependencies": {
     "tslib": "^2.2.0"
